+++ pylib/serial/literate.py
@@ -85,7 +85,7 @@ def write(model, out=sys.stdout, base=None, propertybase=None, shorteners=None):
     if any((base, propertybase, shorteners)):
         out.write('# @docheader\n\n* @iri:\n')
     if base:
-        out.write('    * @base: {0}'.format(base))
+        out.write('    * @nodebase: {0}'.format(base))
     #for k, v in shorteners:
     #    out.write('    * @base: {0}'.format(base))